Separated warzone from admin page.
Using a slug for modularity since wz and wz2 are so similar for kitbases and for future health of adding new games. Need to implement dashboard routing still. Create game button does not work. Add a "Back to admin home page" button.
